Move Google captcha validation to own file
